chore(react): improve react e2e tests speed
