+++ .github/workflows/e2e-windows.yml
@@ -87,7 +87,8 @@ jobs:
           - e2e-nx-misc
           - e2e-nx-plugin
           - e2e-nx-run
-          - e2e-react
+          - e2e-react-core
+          - e2e-react-extensions
           - e2e-web
           - e2e-rollup
           - e2e-storybook
@@ -126,7 +127,9 @@ jobs:
             codeowners: 'S04SYHYKGNP'
           - project: e2e-nx-run
             codeowners: 'S04SYHYKGNP'
-          - project: e2e-react
+          - project: e2e-react-core
+            codeowners: 'S04TNCNJG5N'
+          - project: e2e-react-extensions
             codeowners: 'S04TNCNJG5N'
           - project: e2e-web
             codeowners: 'S04SJ6PL98X'
@@ -173,7 +176,9 @@ jobs:
           - node_version: 16
             project: e2e-lerna-smoke-tests
           - node_version: 16
-            project: e2e-react
+            project: e2e-react-core
+          - node_version: 16
+            project: e2e-react-extensions
           - node_version: 16
             project: e2e-web
           - node_version: 16
@@ -213,7 +218,9 @@ jobs:
           - node_version: 19
             project: e2e-lerna-smoke-tests
           - node_version: 19
-            project: e2e-react
+            project: e2e-react-core
+          - node_version: 19
+            project: e2e-react-extensions
           - node_version: 19
             project: e2e-web
           - node_version: 19
