+++ e2e/react-core/tsconfig.json
