+++ e2e/react-core/tsconfig.spec.json
