+++ e2e/react-extensions/project.json
@@ -1,0 +1,11 @@
+{
+  "name": "e2e-react-extensions",
+  "$schema": "../../node_modules/nx/schemas/project-schema.json",
+  "sourceRoot": "e2e/react-extensions",
+  "projectType": "application",
+  "targets": {
+    "e2e": {},
+    "run-e2e-tests": {}
+  },
+  "implicitDependencies": ["react"]
+}
