+++ e2e/react-extensions/src/cypress-component-tests.test.ts
